Relocate stuff
* Internal package with non-client data structures and classes
* Codecs separate from data structures
