+++ src/main/scala/com/vroste/adsclient/internal/AdsResponse.scala
@@ -1,4 +1,4 @@
-package com.vroste.adsclient
+package com.vroste.adsclient.internal
 
 import java.time.Instant
