+++ src/main/scala/com/vroste/adsclient/internal/codecs/AdsResponseCodecs.scala
@@ -1,11 +1,11 @@
-package com.vroste.adsclient
+package com.vroste.adsclient.internal.codecs
 
 import java.time.temporal.ChronoUnit
 import java.time.{Duration, Instant}
 
-import scodec.codecs.{bytes, listOfN, uint32L, variableSizeBytesLong}
+import com.vroste.adsclient.internal._
+import scodec.codecs.{StringEnrichedWithCodecContextSupport, bytes, listOfN, uint32L, variableSizeBytesLong}
 import scodec.{Attempt, Codec, Err}
-import scodec.codecs.StringEnrichedWithCodecContextSupport
 
 trait AdsResponseCodecs {
   import AdsResponse._
@@ -65,17 +65,3 @@ trait AdsResponseCodecs {
 }
 
 object AdsResponseCodecs extends AdsResponseCodecs
-
-object WindowsFiletime {
-  lazy val timestampZero: Instant = Instant.parse("1601-01-01T00:00:00Z")
-
-  def filetimeToInstant(fileTime: Long): Instant = {
-    val duration = Duration.of(fileTime / 10, ChronoUnit.MICROS).plus(fileTime % 10 * 100, ChronoUnit.NANOS)
-    timestampZero.plus(duration)
-  }
-
-  def instantToFiletime(instant: Instant): Long =
-    timestampZero.until(instant, ChronoUnit.NANOS) // Something like this
-
-  val codec: Codec[Instant] = scodec.codecs.longL(64).xmap(filetimeToInstant, instantToFiletime)
-}
